Get a java classpath from POM

This function requires a simple pom.xml to function, so its usability is limited. However it can be very handy for packages with java dependencies where there is not a fat jar to depend upon. If you have a list of dependencies but no actual pom.xml, see get_classpath

get_classpath_from_pom(path, mvn = find_mvn(), java_home)

Arguments

path
Character. Path to the directory containing a pom.xml to derive paths for

mvn
Character. The path the the maven installation.

java_home
Character. Path to java. If not provided the standard install paths (platform dependent) will be checked.