Write a basic pom.xml

The purpose of this function is to write a pom.xml to support classpath management when using rJava on dependencies with dependencies of their own. For the most part this is probably not a function that users should be calling directly.

write_pom(dep, group, version, mvn = find_mvn(), java_home,
  transitive = TRUE, path = tempdir(), artifactid = "tempproj",
  groupid = "tempgroup", artifactversion = "0.0.1-SNAPSHOT")

Arguments

path
Character. The path where the pom should be written.

artifactid
Character. The artifactId entry in the resulting pom.xml. Unless you intend to save this pom for future use this entry does not really matter and the default is fine.

groupid
Character. The groupId entry in the resulting pom.xml. See artifactid for use case.

artifactversion
Character. The version entry in the resulting pom.xml. See artifactid for use case.

Value

the file path of the newly written pom.xml